+++ packages/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.1-prerelease.10](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@5.1.1-prerelease.9...demo-app-ts@5.1.1-prerelease.10) (2023-09-13)
+
+**Note:** Version bump only for package demo-app-ts
+
+## [5.1.1-prerelease.9](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@5.1.1-prerelease.8...demo-app-ts@5.1.1-prerelease.9) (2023-09-13)
+
+**Note:** Version bump only for package demo-app-ts
+
+## [5.1.1-prerelease.8](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@5.1.1-prerelease.7...demo-app-ts@5.1.1-prerelease.8) (2023-09-12)
+
+**Note:** Version bump only for package demo-app-ts
+
+## 5.1.1-prerelease.7 (2023-09-12)
+
+**Note:** Version bump only for package demo-app-ts
+
 ## [5.1.1-prerelease.6](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@5.1.1-prerelease.5...demo-app-ts@5.1.1-prerelease.6) (2023-09-07)
 
 **Note:** Version bump only for package demo-app-ts